Revert "qcacld-3.0: Enable auto PS timer in hdd config file"
This reverts Change-Id: I2668b7c1ba64b67bf524e449e96ec6d79a2c1676.
Customer do not want to take this fix, as this is a ini change
and with the default value they have not seen the issue on
any device.
They are comfortable not taking this change for now,
hence revert.

Change-Id: I29b0f3b6e3acf25b8d82cfb85d4b1dbd5961336d
CRs-Fixed: 2214124