qcacld-3.0: Send Peer Authorized for SAP peers after Set Key Complete
The hostapd sends the add_key command followed by the change_station
command to set the peer to authorized state and open the data ports.
Since the add_key is serialized through LIM/WMA layers and might not
have been sent to the FW by the time the change_station is received,
the FW would assert due to peer moved to authorized state before keys
are set.

Send the Peer authorized command to FW and to the hostapd only after
the set_key is complete for the peer.

Change-Id: Ibd1737a41ea96770987da4070b72e756538e7508
CRs-Fixed: 2267291
Signed-off-by: hsuvictor <hsuvictor@google.com>