qcacld-3.0: Check TDLS peer type in lim_process_add_sta_rsp
Currenly, as part of tdls add sta req gLimAddStaTdls is set to 1.
And as part of add sta response gLimAddStaTdls is checked if 1 to
map tdls add sta response.
This implementation needs to take care of all error cases of
add sta req failure to reset the gLimAddStaTdls value to 0.

Instead check for peer type in add sta rsp.

Change-Id: Ideaff239f743b95a9578806d2ec220e123d4d995
CRs-Fixed: 2281385
Bug: 112294368
Signed-off-by: Srinivas Girigowda <sgirigow@codeaurora.org>