techpack/audio: b4s4: ASoC: correct the TDM probe
The probe process of msm_dai_q6_probe and msm_dai_q6_tdm_dev_probe are
not definitely executed in pairs, and the current cache mechanism should
get failed if the sequence of probe DAI and probe TDM device is not in
the expected order. In this case, msm_dai_q6_tdm_dev_probe is not
executed at the first time because the LPI has not been probed yet and
the corresponding configs are not set correctly.

Bug: 117047530
Test: boot to home/VoIP(Skype)/Phone call
Change-Id: I733f2d846ed72baafea5ee806d215aeccb697008
Signed-off-by: HW Lee <hwlee@google.com>