asoc: wcd-dsp-mgr: retry firmware image download
If image download for the DSP fails, it is possible that the
driver always stays in bad state and will reject any future
requests to enable the DSP. Add change to manage the driver
state such that image download can be re-attempted each time
there is request for DSP enablement. While at it, move the
mutex lock in wdsp_enable_dsp to make sure the state check
is performed after the ssr_lock is acquired to avoid race
of the state variable between SSR and DSP enable.

This also address the wakeup source not initialized issue

Bug: 119823747
Test: manual
Change-Id: Id29ce918ee125126670612652d3a5f329ca46933
Signed-off-by: Bhalchandra Gajare <gajare@codeaurora.org>
Signed-off-by: Carter Hsu <carterhsu@google.com>