msm: kgsl: Get names of all fences
Sometimes, we may be given a fence array to wait on.
And fence array goes by a generic name, so it is hard
to tell to which driver the individual fences belong
to. Store names of all individual fences that are in
the fence array so that we can print them in case of
fence timeouts or tracepoints for better debugging.

Change-Id: I466606a5c10db11a5737fc1014185a9e72df5391
Signed-off-by: Harshdeep Dhatt <hdhatt@codeaurora.org>
Signed-off-by: Puranam V G Tejaswi <pvgtejas@codeaurora.org>
(cherry picked from commit 0ebbdcdc1fa10b8ca31feac93fae237cfa0eb8c5)