soc: wcd-spi-ac: fix spi access handling
It is observed in testing that sometimes the QMI response to indicate
that the bus access is released times out. This causes the driver state
to indicate the spi bus access is local. But if the remote side has
already received the QMI message then it is possible that SPI bus is
being accessed by the remote processor as well as the local processor
leading the SPI bus into a bad state. Fix the handling such that even if
QMI message times out, access is still considered as remote. While at
it, increase the QMI timeout to 1 sec from 500msec and add retry logic
for the request access message.

Bug: 119823747
Test: manual
Change-Id: I209b3906ab6fdb32319fbd889d6fbd66697107ad
Signed-off-by: Bhalchandra Gajare <gajare@codeaurora.org>