qcacld-3.0: Process ll stats correctly
Currently in the api hdd_process_ll_stats, if the stats
type is iface stats in that case the mem copy to the final
result uses the size of the structure pointer instead of
size of the structure which results in incorrect value for
the iface stats.

To address above issue use the size of the structure for
the memcopy of the iface stats.

Bug: 158521704
CRs-Fixed: 2710446
Change-Id: I2b6128f80fa994a5d2ed2fdd2feeaa1ae2e341d3
(cherry picked from commit 8450514e45e4a4a8a74a451ab06b58d19561d56b)