qcacld-3.0: Update rate flags correctly
Currently in driver does  not update the rate flags correctly
in wma as rate flags should include all the subsets of the
lower rartesets, which is not thye case today and driver only
updates the higher rate flag. Because of which it leads to
invalid computation of txrate at the kernel.

Bug: 166722837
Change-Id: I5529532b3d41b68693b5b4b8952ee0f1414354db
CRs-Fixed: 2776370